src/LeaseConversations/requirements.txt
flask
openai
boto3==1.26.0
pdfminer.six==20231228
python-dotenv
